requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
itsdangerous==2.2.0
docker==7.1.0
python-dotenv==1.0.1